+++ inc/settings-page-admin.php
@@ -216,8 +216,9 @@ function elodin_bridge_render_admin_page() {
 	$edit_site_admin_bar_links_enabled = elodin_bridge_is_edit_site_admin_bar_links_enabled();
 	$site_editor_admin_bar_enabled = elodin_bridge_is_site_editor_admin_bar_enabled();
 	$mobile_fixed_background_repair_enabled = elodin_bridge_is_mobile_fixed_background_repair_enabled();
-	$sticky_below_header_block_style_settings = elodin_bridge_get_sticky_below_header_block_style_settings();
+	$header_aware_positioning_settings = elodin_bridge_get_header_aware_positioning_settings();
 	$sticky_below_header_block_style_enabled = elodin_bridge_is_sticky_below_header_block_style_enabled();
+	$anchor_smooth_scroll_enabled = elodin_bridge_is_anchor_smooth_scroll_enabled();
 	$scroll_hide_elements_settings = elodin_bridge_get_scroll_hide_elements_settings();
 	$scroll_hide_elements_enabled = elodin_bridge_is_scroll_hide_elements_enabled();
 	$scroll_hide_element_rows = array_values( $scroll_hide_elements_settings['rules'] );